Fix: 건물 상세 정보 및 infra 정보 구현 & auto_increment 삭제, 배치 정보 수정

+++ Springboot/houscore/src/main/java/com/hs/houscore/dto/RecommendAiDTO.java
@@ -9,7 +9,7 @@ public class RecommendAiDTO {
     private String address;
     private Double aiScore;
     private Double realPrice;   //만원 단위
-    private Double pricePerPyeong;
+    private Integer pricePerPyeong;
     private Double pricePerRegion ; //지역대비가격 %
     private Long reviewCnt;
 }
